lazy: Set missing install colorscheme to default
Previously, the default was to use habamax which sets terminal colors
(see help for `terminal-config`). Currently, molokai is used which does
not set terminal colors since it is very old and predates terminal mode.

The result was that after a missing plugin install on startup, uses of
terminal mode (such as lazygit through the floaterm plugin) would stay
using the habamax colors.

This is fixed by using the default colorscheme on startup install, since
it does not set any terminal colors.